+++ source_gen/source_gen/source_gen/CHANGELOG.md
@@ -1,3 +1,5 @@
+## 1.2.8-dev
+
 ## 1.2.7
 
 * Update the value of the pubspec `repository` field.